Fix format specifier

+++ src/fs.c
@@ -1,4 +1,5 @@
 #include <fcntl.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <string.h>
 #include <sys/stat.h>
@@ -94,7 +95,7 @@ FILE *open_file(const char *const restrict path, const char *const restrict mode
     struct stat st;
 
     if (fstat(fd, &st) == 0) {
-        logger_printf(LOG_LEVEL_DEBUG, "%s [%s] (%lu bytes)\n", path, mode, st.st_size);
+        logger_printf(LOG_LEVEL_DEBUG, "%s [%s] (%" PRIdMAX " bytes)\n", path, mode, (intmax_t) st.st_size);
     } else {
         logger_perror("fstatat");
     }
